Conflicts:
	CHANGE_LOG.TXT
	docs/html/annotated.html
	docs/html/block__discontinuity_8cuh.html
	docs/html/block__exchange_8cuh.html
	docs/html/block__load_8cuh.html
	docs/html/block__radix__sort_8cuh.html
	docs/html/block__reduce_8cuh.html
	docs/html/block__scan_8cuh.html
	docs/html/block__store_8cuh.html
	docs/html/citelist.html
	docs/html/classcub_1_1_block_discontinuity-members.html
	docs/html/classcub_1_1_block_discontinuity.html
	docs/html/classcub_1_1_block_exchange-members.html
	docs/html/classcub_1_1_block_exchange.html
	docs/html/classcub_1_1_block_load-members.html
	docs/html/classcub_1_1_block_load.html
	docs/html/classcub_1_1_block_radix_sort-members.html
	docs/html/classcub_1_1_block_radix_sort.html
	docs/html/classcub_1_1_block_reduce-members.html
	docs/html/classcub_1_1_block_reduce.html
	docs/html/classcub_1_1_block_scan-members.html
	docs/html/classcub_1_1_block_scan.html
	docs/html/classcub_1_1_block_store-members.html
	docs/html/classcub_1_1_block_store.html
	docs/html/classcub_1_1_warp_reduce-members.html
	docs/html/classcub_1_1_warp_reduce.html
	docs/html/classcub_1_1_warp_scan-members.html
	docs/html/classcub_1_1_warp_scan.html
	docs/html/classes.html
	docs/html/dir_011e1c944d88f71be72e1e24a5fda7cf.html
	docs/html/dir_bb50a5ef59f19d030d06415663184d05.html
	docs/html/dir_cb3a671affffe7eeb3fdf5ae58e42cc8.html
	docs/html/dir_d583f216f1aafe19404e836b0c097ad2.html
	docs/html/functions.html
	docs/html/functions_func.html
	docs/html/functions_rela.html
	docs/html/functions_type.html
	docs/html/functions_vars.html
	docs/html/group___block_module.html
	docs/html/group___device_module.html
	docs/html/group___simt_coop.html
	docs/html/hierarchy.html
	docs/html/index.html
	docs/html/modules.html
	docs/html/namespacecub.html
	docs/html/structcub_1_1_base_traits-members.html
	docs/html/structcub_1_1_base_traits.html
	docs/html/structcub_1_1_device_reduce-members.html
	docs/html/structcub_1_1_enable_if-members.html
	docs/html/structcub_1_1_enable_if.html
	docs/html/structcub_1_1_equality-members.html
	docs/html/structcub_1_1_equality.html
	docs/html/structcub_1_1_equals-members.html
	docs/html/structcub_1_1_equals.html
	docs/html/structcub_1_1_if-members.html
	docs/html/structcub_1_1_if.html
	docs/html/structcub_1_1_is_volatile-members.html
	docs/html/structcub_1_1_is_volatile.html
	docs/html/structcub_1_1_log2-members.html
	docs/html/structcub_1_1_log2.html
	docs/html/structcub_1_1_max-members.html
	docs/html/structcub_1_1_max.html
	docs/html/structcub_1_1_null_type.html
	docs/html/structcub_1_1_numeric_traits-members.html
	docs/html/structcub_1_1_numeric_traits.html
	docs/html/structcub_1_1_remove_qualifiers-members.html
	docs/html/structcub_1_1_remove_qualifiers.html
	docs/html/structcub_1_1_sum-members.html
	docs/html/structcub_1_1_sum.html
	docs/html/structcub_1_1_traits-members.html
	docs/html/structcub_1_1_traits.html
	docs/html/thread__load_8cuh.html
	docs/html/thread__operators_8cuh.html
	docs/html/thread__store_8cuh.html
	docs/html/util__debug_8cuh.html
	docs/html/util__device_8cuh.html
	docs/html/util__type_8cuh.html
	docs/html/warp__reduce_8cuh.html
	docs/html/warp__scan_8cuh.html

+++ cub/util_allocator.cuh
@@ -341,7 +341,7 @@ struct CachingDeviceAllocator : public DeviceAllocator
 
         this->max_cached_bytes = max_cached_bytes;
 
-        if (debug) printf("New max_cached_bytes(%d)\n", max_cached_bytes);
+        if (debug) printf("New max_cached_bytes(%lld)\n", (long long) max_cached_bytes);
 
         // Unlock
         Unlock(&spin_lock);
@@ -400,8 +400,8 @@ struct CachingDeviceAllocator : public DeviceAllocator
                 cached_blocks.erase(block_itr);
                 cached_bytes[gpu] -= search_key.bytes;
 
-                if (debug) printf("\tGPU %d reused cached block (%d bytes). %d available blocks cached (%d bytes), %d live blocks outstanding.\n",
-                    gpu, search_key.bytes, cached_blocks.size(), cached_bytes[gpu], live_blocks.size());
+                if (debug) printf("\tGPU %d reused cached block (%lld bytes). %lld available blocks cached (%lld bytes), %lld live blocks outstanding.\n",
+                    gpu, (long long) search_key.bytes, (long long) cached_blocks.size(), (long long) cached_bytes[gpu], (long long) live_blocks.size());
             }
             else
             {
@@ -427,8 +427,8 @@ struct CachingDeviceAllocator : public DeviceAllocator
                 // Insert into live blocks
                 live_blocks.insert(search_key);
 
-                if (debug) printf("\tGPU %d allocating new device block %d bytes. %d available blocks cached (%d bytes), %d live blocks outstanding.\n",
-                    gpu, search_key.bytes, cached_blocks.size(), cached_bytes[gpu], live_blocks.size());
+                if (debug) printf("\tGPU %d allocating new device block %lld bytes. %lld available blocks cached (%lld bytes), %lld live blocks outstanding.\n",
+                    gpu, (long long) search_key.bytes, (long long) cached_blocks.size(), (long long) cached_bytes[gpu], (long long) live_blocks.size());
             }
         } while(0);
 
@@ -507,8 +507,8 @@ struct CachingDeviceAllocator : public DeviceAllocator
                     cached_blocks.insert(search_key);
                     cached_bytes[gpu] += search_key.bytes;
 
-                    if (debug) printf("\tGPU %d returned %d bytes. %d available blocks cached (%d bytes), %d live blocks outstanding.\n",
-                        gpu, search_key.bytes, cached_blocks.size(), cached_bytes[gpu], live_blocks.size());
+                    if (debug) printf("\tGPU %d returned %lld bytes. %lld available blocks cached (%lld bytes), %lld live blocks outstanding.\n",
+                        gpu, (long long) search_key.bytes, (long long) cached_blocks.size(), (long long) cached_bytes[gpu], (long long) live_blocks.size());
                 }
                 else
                 {
@@ -525,8 +525,8 @@ struct CachingDeviceAllocator : public DeviceAllocator
                     // Free device memory
                     if (CubDebug(error = cudaFree(d_ptr))) break;
 
-                    if (debug) printf("\tGPU %d freed %d bytes.  %d available blocks cached (%d bytes), %d live blocks outstanding.\n",
-                        gpu, search_key.bytes, cached_blocks.size(), cached_bytes[gpu], live_blocks.size());
+                    if (debug) printf("\tGPU %d freed %lld bytes.  %lld available blocks cached (%lld bytes), %lld live blocks outstanding.\n",
+                        gpu, (long long) search_key.bytes, (long long) cached_blocks.size(), (long long) cached_bytes[gpu], (long long) live_blocks.size());
                 }
             }
         } while (0);
@@ -606,8 +606,8 @@ struct CachingDeviceAllocator : public DeviceAllocator
             cached_bytes[current_gpu] -= begin->bytes;
             cached_blocks.erase(begin);
 
-            if (debug) printf("\tGPU %d freed %d bytes.  %d available blocks cached (%d bytes), %d live blocks outstanding.\n",
-                current_gpu, begin->bytes, cached_blocks.size(), cached_bytes[current_gpu], live_blocks.size());
+            if (debug) printf("\tGPU %d freed %lld bytes.  %lld available blocks cached (%lld bytes), %lld live blocks outstanding.\n",
+                current_gpu, (long long) begin->bytes, (long long) cached_blocks.size(), (long long) cached_bytes[current_gpu], (long long) live_blocks.size());
         }
 
         // Unlock
